[modify] 화면 UI 수정
