+++ pp/pp/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0x70",
-          "green" : "0x0B",
-          "red" : "0x00"
+          "blue" : "0x33",
+          "green" : "0xD2",
+          "red" : "0xFF"
         }
       },
       "idiom" : "universal"
