+++ pp/pp/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "appico.png",
+      "filename" : "app.png",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
