+++ pp/pp/Assets.xcassets/no.post.imageset/Contents.json
@@ -1,17 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "Group 34341.png",
+      "filename" : "post-it note with pin.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "Group 34341@2x.png",
+      "filename" : "post-it note with pin@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "Group 34341@3x.png",
+      "filename" : "post-it note with pin@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
