+++ pp/pp/View/Community/CommunityDetailView.swift
@@ -15,79 +15,81 @@ struct CommunityDetailView: View {
   
     @State private var showAlert = false
     @State private var showReportConfirmation = false  // 신고 처리 확인용 Alert 표시
- 
-    var body: some View {
-        VStack(alignment: .leading) {
-
-            if let imageUrls = vm.postDetail?.imageUrls, !imageUrls.isEmpty {
-                           AutoScroller2(imageURLs: imageUrls)
-                               .frame(height: 258)
-                       } else {
-                           ProgressView()
-                               .frame(height: 258)
-                       }
-            
-            Text(vm.postDetail?.title ?? "title")
-                .font(.system(size: 18))
-                .padding(.top, 25)
-            
-            Text(vm.postDetail?.createdDate ?? "date")
-                .font(.system(size: 12))
-            
-            Text(vm.postDetail?.content ?? "content")
-                .font(.system(size: 15))
-                .padding(.top, 20)
-            
-            LikeAndReplyView(vm: vm,postId:postId)
-            Spacer()
-        }
-        .onAppear {
-            print("Loading post details for postId: \(postId)")
-                   vm.loadDetailPosts(postId: postId)
-               }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 25)
-        
-        .toolbar {
-            ToolbarItem {
-                Menu {
-                    Button(role: .destructive) {
-                        showAlert = true
-                    } label: {
-                      Label("신고", systemImage: "exclamationmark.circle")
-                            .frame(width: 22, height: 30)
-                    }
-                } label: {
-                    Image("menu.icon")
-                        .frame(width: 22, height: 30)
-                }
-            }
-        }
-        .alert("신고 확인", isPresented: $showAlert) {
-            Button("확인", role: .destructive) {
-                showReportConfirmation = true
-                
-                // ToDo: - fetchPost에서 각 게시글 Id 받아와서 가지고 있다 신고할때 이 Id 값으로 신고
-                vm.reportPost(postId: self.postId)
-                print("신고 postId\(postId)")
-            }
-            Button("취소", role: .cancel) {}
-        } message: {
-            Text("이 게시물을 신고하시겠습니까?")
-        }
-        .alert("신고 완료", isPresented: $showReportConfirmation) {
-            Button("확인", role: .cancel) {
-                dismiss()
-            }
-        } message: {
-            Text("신고가 처리되었습니다.")
-        }
+	
+	var body: some View {
+		GeometryReader { geometry in
+			VStack(alignment: .leading) {
+				if let imageUrls = vm.postDetail?.imageUrls, !imageUrls.isEmpty {
+					AutoScroller2(imageURLs: imageUrls, size: abs(geometry.size.width - 32))
+						.frame(width: abs(geometry.size.width - 32), height: abs(geometry.size.width - 32))
+				} else {
+					ProgressView()
+						.frame(width: abs(geometry.size.width - 32), height: abs(geometry.size.width - 32))
+				}
+				
+				Text(vm.postDetail?.title ?? "title")
+					.font(.system(size: 18))
+					.padding(.top, 25)
+				
+				Text(vm.postDetail?.createdDate ?? "date")
+					.font(.system(size: 12))
+				
+				Text(vm.postDetail?.content ?? "content")
+					.font(.system(size: 15))
+					.padding(.top, 20)
+				
+				LikeAndReplyView(vm: vm,postId:postId)
+				Spacer()
+			}
+			.onAppear {
+				print("Loading post details for postId: \(postId)")
+				vm.loadDetailPosts(postId: postId)
+			}
+			.padding(.horizontal, 16)
+			.padding(.vertical, 25)
+			
+			.toolbar {
+				ToolbarItem {
+					Menu {
+						Button(role: .destructive) {
+							showAlert = true
+						} label: {
+							Label("신고", systemImage: "exclamationmark.circle")
+								.frame(width: 22, height: 30)
+						}
+					} label: {
+						Image("menu.icon")
+							.frame(width: 22, height: 30)
+					}
+				}
+			}
+			.alert("신고 확인", isPresented: $showAlert) {
+				Button("확인", role: .destructive) {
+					showReportConfirmation = true
+					
+						// ToDo: - fetchPost에서 각 게시글 Id 받아와서 가지고 있다 신고할때 이 Id 값으로 신고
+					vm.reportPost(postId: self.postId)
+					print("신고 postId\(postId)")
+				}
+				Button("취소", role: .cancel) {}
+			} message: {
+				Text("이 게시물을 신고하시겠습니까?")
+			}
+			.alert("신고 완료", isPresented: $showReportConfirmation) {
+				Button("확인", role: .cancel) {
+					dismiss()
+				}
+			} message: {
+				Text("신고가 처리되었습니다.")
+			}
+		}
     }
 }
 
 
 struct AutoScroller2: View {
     var imageURLs: [URL]
+	let size: CGFloat
 
     var body: some View {
         ZStack {
@@ -100,16 +102,20 @@ struct AutoScroller2: View {
                         switch phase {
                         case .empty:
                             ProgressView()
+								.frame(width: size, height: size)
+								.clipShape(RoundedRectangle(cornerRadius: 10))
                         case .success(let image):
                             image
                                 .resizable()
-                                .scaledToFit()
-                                .frame(height: 258)
+                                .scaledToFill()
+								.frame(width: size, height: size)
+								.clipShape(RoundedRectangle(cornerRadius: 10))
                         case .failure:
-                            Image(systemName: "photo")
+                            Image(systemName: "empty.image")
                                 .resizable()
-                                .scaledToFit()
-                                .frame(height: 258)
+                                .scaledToFill()
+								.frame(width: size, height: size)
+								.clipShape(RoundedRectangle(cornerRadius: 10))
                         @unknown default:
                             EmptyView()
                         }
@@ -142,6 +148,7 @@ struct LikeAndReplyView: View {
             } label: {
                 HStack {
                     Text("좋아요")
+						.tint(.black)
                     Image(systemName: vm.isLiked ? "heart.fill" : "heart")
                         .foregroundColor(vm.isLiked ? .red : .black)
                 }
@@ -151,7 +158,9 @@ struct LikeAndReplyView: View {
             NavigationLink(destination: PostReplyView(vm: vm,postId:postId)) {
                 HStack {
                     Text("댓글")
+						.tint(.black)
                     Image(systemName: "bubble")
+						.tint(.black)
                 }
             }
             Text("\(vm.commentCounts)")
